online_inference/requirements.txt
fastapi[uvicorn]~=0.88.0
python-dotenv~=0.21.0
fastapi-health==0.4.0
pydantic==1.10.2
uvicorn==0.20.0
scikit-learn==1.1.3
typing==3.7.4.3
pandas~=1.5.2
